RecipeContext är nu tillagd som en Service
