+++ OGFoodAPI/Controllers/ApiRequestTest.cs
@@ -11,6 +11,13 @@ namespace OGFoodAPI.Controllers
     [ApiController]
     public class ApiRequestTest : ControllerBase
     {
+        readonly IRecipeContext _recipeContext;
+
+        public ApiRequestTest(IRecipeContext recipeContext)
+        {
+            _recipeContext = recipeContext;
+        }
+
         [HttpGet("{search}")]
         [ProducesResponseType(StatusCodes.Status200OK)]
         public async Task<string> Index(string search)
@@ -18,18 +25,21 @@ namespace OGFoodAPI.Controllers
             if (search == null)
                 return "No data to show";
 
-            var ingredient1 = new IngredientWithAmount() { Amount = 1, Ingredient = new Ingredient() { Name = "Lök", Id="1" } };
-            var ingredient2 = new IngredientWithAmount() { Amount = 1, Ingredient = new Ingredient() { Name = "Tomatpure", Id="3" } };
+            //Det här är endast för att ge  exempel på en söksträng man kan använda
+            var ingredient1 = new IngredientWithAmount() { Amount = 1, Ingredient = new Ingredient() { Id = "1", Name = "Lök" } };
+            var ingredient2 = new IngredientWithAmount() { Amount = 1, Ingredient = new Ingredient() { Id = "2", Name = "Falukorv" } };
+            var ingredient3 = new IngredientWithAmount() { Amount = 2, Ingredient = new Ingredient() { Id = "3", Name = "Tomatpuré" } };
             var ingredientList = new List<IngredientWithAmount>();
             ingredientList.Add(ingredient1);
             ingredientList.Add(ingredient2);
+            ingredientList.Add(ingredient3);
 
             var reqJson = JsonConvert.SerializeObject(ingredientList);
 
             Console.WriteLine("example request: " + reqJson);
+            //----------------------------------------------------------------------
 
-            var searchLocalStorage = new InitializeLocalStorage();
-            return await searchLocalStorage.GetRecipes(search);
+            return await _recipeContext.GetRecipes(search);
 
             throw new Exception("Error requesting data");
         }
