+++ OGFoodAPI/RecipeService/RecipeContext.cs
@@ -15,6 +15,11 @@ namespace OGFoodAPI.RecipeService
             return _apiCaller.DeserializeData(data, apiRequest);
         }
 
+        public Task<string> GetRecipes(string search)
+        {
+            return _apiCaller.GetRecipes(search);
+        }
+
         public async Task<ApiResponse> Request(ApiRequest apiRequest)
         {
             return await _apiCaller.Request(apiRequest);
